+++ src/gui/components/navbar/navbar.spec.js
