+++ src/gui/screens/workbench/props/input/input.styles.js
