+++ src/gui/screens/workbench/props/props.spec.js
