+++ src/gui/screens/workbench/props/props.styles.js
