+++ src/gui/screens/workbench/workbench.styles.js
